# be/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Kredius API
  description: |
    Kredius Backend REST API — OpenAPI-first design.

    ## How it works
    This YAML file is the **single source of truth** for all API endpoints and model classes.
    The `openapi-generator-maven-plugin` reads this file during the `generate-sources` phase and
    produces:
    - `com.kredius.be.api.*` — Spring MVC interfaces annotated with `@RequestMapping`
    - `com.kredius.be.model.*` — Kotlin data classes for request/response bodies

    ## Adding a new endpoint
    1. Define the path + operation in this file
    2. Run `mvn generate-sources`
    3. Implement the updated interface in the matching `controller/` class

  version: 1.0.0
  contact:
    name: Kredius Engineering
    email: dev@kredius.com

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Health
    description: Health and readiness checks
  - name: Accounts
    description: Chart of accounts management
  - name: Loans
    description: Loan management

paths:

  /health:
    get:
      tags: [Health]
      summary: Health check
      description: Returns the health status of the API
      operationId: healthCheck
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /api/v1/accounts/summary:
    get:
      tags: [Accounts]
      summary: Account summary list
      description: Returns active accounts with current balance and last transaction for the current user
      operationId: getAccountsSummary
      responses:
        '200':
          description: Account summaries retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountSummaryResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/v1/accounts:
    get:
      tags: [Accounts]
      summary: List active accounts
      description: Returns all active accounts for the current user, optionally filtered by type
      operationId: getAccounts
      parameters:
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AccountType'
      responses:
        '200':
          description: Accounts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Accounts]
      summary: Create an account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/v1/loans:
    get:
      tags: [Loans]
      summary: List loans
      description: Returns all active loans for the current user, optionally filtered by type
      operationId: getLoans
      parameters:
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/LoanType'
      responses:
        '200':
          description: Loans retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoanResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  schemas:

    AccountSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        code:
          type: string
          nullable: true
        name:
          type: string
        type:
          $ref: '#/components/schemas/AccountType'
        showInAlerts:
          type: boolean
        loanAccount:
          type: boolean
        balance:
          type: number
          format: double
          description: Signed balance in RD$ (positive = normal balance for account type)
        lastTransactionDate:
          type: string
          format: date
          nullable: true
        lastTransactionDescription:
          type: string
          nullable: true
        lastTransactionValue:
          type: number
          format: double
          nullable: true
          description: RD$ amount of the last journal line for this account
        thresholdPct:
          type: number
          format: double
          nullable: true
          description: Budget threshold as a percentage of income (e.g. 20 = 20%)
        loanId:
          type: integer
          format: int64
          nullable: true
          description: Present only when loanAccount is true
        loanType:
          $ref: '#/components/schemas/LoanType'
          nullable: true
        loanCounterpartyName:
          type: string
          nullable: true
        loanInstallmentAmount:
          type: number
          format: double
          nullable: true
        loanNextInstallmentDate:
          type: string
          format: date
          nullable: true
        loanRemainingBalance:
          type: number
          format: double
          nullable: true
        loanTotalAmount:
          type: number
          format: double
          nullable: true

    AccountType:
      type: string
      enum:
        - ASSET
        - LIABILITY
        - EQUITY
        - INCOME
        - EXPENSE

    AccountResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        code:
          type: string
          example: '1020'
          nullable: true
        name:
          type: string
          example: Cuenta Ahorros BHD
        type:
          $ref: '#/components/schemas/AccountType'
        thresholdPct:
          type: number
          format: double
          nullable: true
          example: 20.00
        showInAlerts:
          type: boolean
        active:
          type: boolean
        loanAccount:
          type: boolean
          description: True when this account is managed by a Loan record — shown in the Loans section, not the regular account list
        createdAt:
          type: string
          format: date-time
          readOnly: true

    CreateAccountRequest:
      type: object
      required:
        - name
        - type
      properties:
        code:
          type: string
          maxLength: 10
          nullable: true
        name:
          type: string
          maxLength: 120
        type:
          $ref: '#/components/schemas/AccountType'
        thresholdPct:
          type: number
          format: double
          nullable: true
        showInAlerts:
          type: boolean
          default: false

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: UP
        timestamp:
          type: string
          format: date-time

    LoanType:
      type: string
      enum: [GIVEN, RECEIVED]

    LoanFrequency:
      type: string
      enum: [WEEKLY, MONTHLY]

    LoanResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        type:
          $ref: '#/components/schemas/LoanType'
        counterpartyName:
          type: string
        accountCode:
          type: string
          nullable: true
        accountName:
          type: string
        principal:
          type: number
          format: double
        installmentAmount:
          type: number
          format: double
        frequency:
          $ref: '#/components/schemas/LoanFrequency'
        totalInstallments:
          type: integer
          nullable: true
        paidInstallments:
          type: integer
        nextInstallmentDate:
          type: string
          format: date
          nullable: true
        remainingBalance:
          type: number
          format: double
        totalAmount:
          type: number
          format: double
        active:
          type: boolean

    CreateLoanRequest:
      type: object
      required:
        - amount
        - term
      properties:
        amount:
          type: number
          format: double
          minimum: 100
          maximum: 1000000
          description: Requested loan amount in USD
          example: 10000.00
        term:
          type: integer
          minimum: 1
          maximum: 360
          description: Loan term in months
          example: 12

    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Request validation failed
        details:
          type: array
          items:
            type: string

  responses:
    BadRequest:
      description: Bad request — validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'